aws-local-python-datastream/tests/local_producer/requirements.txt
kafka-python==2.0.2
Faker==23.0.0
PyYAML==6.0.1
python-dotenv==1.0.0
